+++ test.sh
@@ -25,6 +25,8 @@ MAIN_TESTS=(
     test_detect_challenge
     test_detect_challenge_scroll_into_view
     test_detect_challenge_raw_dom_selectors
+    # test_fingerprint.sh
+    test_fingerprint_consistency
     test_get_interactive_elements
     test_get_resolution
     test_calibrate